refactor(jsx): tenant-manager.jsx decomposition Phase 1 — extract data/styles/utils/sidebar
Issue #153 Phase 1. Pure structural refactor — behavior identical, every
reference preserved by name. Extracts the four lowest-risk slices of the
1691-line tenant-manager.jsx into their own files under a new
`docs/interactive/tools/tenant-manager/` subdirectory:

  77 lines  tenant-manager/fixtures/demo-tenants.js  (DEMO_TENANTS, DEMO_GROUPS)
 410 lines  tenant-manager/styles.js                  (styles object)
  49 lines  tenant-manager/utils/yaml-generators.js   (maintenance + silent YAML)
 140 lines  tenant-manager/components/GroupSidebar.jsx
1100 lines  tenant-manager.jsx                        (was 1691; -35%)

Why these four first
====================

These are the lowest-risk slices because the extracted code has no
closure dependencies on orchestrator state:

  - data is data (static fixtures)
  - styles are styles (no React, no closures)
  - yaml-generators are pure functions (no React, no closures)
  - GroupSidebar takes everything via props (its useState state is
    purely local — the form's expand/collapse and new-group input
    fields, none of which the orchestrator reads back)

The remaining work — extracting `useTenantData` / `useApiClient` /
`useGroups` / `useModalFocusTrap` custom hooks and the render-tree
sub-components (TenantCard / ApiNotificationToast / OverflowBanner /
ActiveFiltersBar) — requires careful prop-wiring + closure-semantics
design (most of those need to share `setApiNotification`, `setSelected`,
`setActiveGroupId`, etc.) and is deferred to PR-2d Phase 2. Issue #153
stays open with Phase 2 scope.

Wiring mechanism
================

Each extracted file self-registers via `window.__X = X;` at its tail,
and the orchestrator pulls each via `const X = window.__X;` near the
top. This is the same pattern self-service-portal uses for its three
tab modules (AlertPreviewTab / YamlValidatorTab / RoutingTraceTab).

It's the only safe way to share names across the dep chain because
jsx-loader evaluates deps via `(0, eval)(code)` (indirect eval), where
`const` and `let` declarations are block-scoped to the eval frame and
do NOT leak to the surrounding global scope. The naive `const X = ...`
pattern would compile but the orchestrator would then see a
`ReferenceError: X is not defined` at runtime. Documented inline above
the `const X = window.__X;` block in tenant-manager.jsx.

Lint scan recursion
===================

`scripts/tools/lint/lint_jsx_babel.py` was scanning only top-level
`*.jsx` in the two JSX_DIRS via `glob("*.jsx")`. The new nested layout
(`tenant-manager/components/*.jsx` etc.) requires recursive walking,
so this commit changes the glob to `rglob("*.jsx")`. All severity
rules from issue #152 apply identically to newly-scanned files. The
existing 34 unit tests still pass.

doc-map regeneration
====================

`docs/internal/doc-map.md` and `doc-map.en.md` regenerated to include
the new `tenant-manager/components/GroupSidebar.jsx` entry. The new
`.js` files (data / styles / utils) aren't tracked because doc-map
indexes only `.md` and `.jsx`. Convention matches how existing
sub-modules (AlertPreviewTab / YamlValidatorTab) are listed.

Verification
============

  - 34/34 lint unit tests pass locally
  - All 4 extracted files have proper YAML front-matter + window
    self-registration
  - Orchestrator references all 6 extracted symbols; no inline
    declarations remain
  - Each file under the 400-line aspirational target except styles.js
    (410 — the styles object is naturally unitary content; splitting
    by concern would just make readers chase imports for no benefit)
