+++ CHANGELOG.md
@@ -70,6 +70,16 @@ Breaking / Upgrade 七塊清楚區分），那是目標形狀。
 
 ### Added
 
+- **`tenant-manager.jsx` decomposition Phase 1（v2.8.0, refactor, issue #153 Phase 1）** — pure structural refactor extracting the four lowest-risk slices of the 1691-line orchestrator into their own files under a new `docs/interactive/tools/tenant-manager/` subdirectory. **Behavior identical** — every reference is preserved by name (no API changes, no JSX changes, no logic changes); the only mechanical change is that `DEMO_TENANTS` / `DEMO_GROUPS` / `styles` / `generateMaintenanceYaml` / `generateSilentModeYaml` / `GroupSidebar` are now loaded as window globals via the front-matter `dependencies: [...]` block (the same pattern self-service-portal uses for its tab modules).
+  - **`tenant-manager/fixtures/demo-tenants.js`** (77 lines) — `DEMO_TENANTS` + `DEMO_GROUPS` static demo fixtures, used by the docs-site fallback path when both the live API and `platform-data.json` are unavailable.
+  - **`tenant-manager/styles.js`** (410 lines) — the styles object mapped to `var(--da-*)` design tokens. Includes inline rationale for the bridge-token pixel literals (6 / 11 / 14 / 18 / 28 / 48px) and component-specific layout constants (280 / 140 / 180 / 600 / 420 / 1400 / 1600px).
+  - **`tenant-manager/utils/yaml-generators.js`** (49 lines) — `generateMaintenanceYaml` + `generateSilentModeYaml` ConfigMap emitters used by the bulk-action modal. Pure string assembly, no React, no closures.
+  - **`tenant-manager/components/GroupSidebar.jsx`** (140 lines) — the group sidebar function component. Stateful only locally (its create-form expand/collapse via own `useState`); accepts props for `groups` + callbacks. ARIA / keyboard nav semantics preserved.
+  - **`tenant-manager.jsx` orchestrator**: 1691 → 1100 lines (~35% reduction). Front-matter now declares `dependencies: [...]` with the 4 helper paths. Top of file pulls each symbol via `const X = window.__X;` immediately after the `import React` line so they're in scope for the rest of the orchestrator. The decomposition strategy comment explains why (`(0, eval)(code)` indirect-eval semantics: `const`/`let` block-scoped to the eval frame, so deps explicitly self-register on `window` at their tail).
+  - **`scripts/tools/lint/lint_jsx_babel.py` updated to `rglob("*.jsx")`** — the lint previously scanned only top-level `*.jsx` in the two `JSX_DIRS`. New nested layout (`tenant-manager/components/*.jsx` etc.) requires recursive walking. Same severity rules apply to all newly-scanned files.
+  - **Honest scope (Phase 1 vs full PR-2d)**: this PR ships the four lowest-risk extractions where extracted code has no closure dependencies on orchestrator state (data is data, styles are styles, YAML generators are pure, GroupSidebar takes everything via props). The remaining work — extracting `useTenantData` / `useApiClient` / `useGroups` / `useModalFocusTrap` custom hooks and the render-tree sub-components (`TenantCard`, `ApiNotificationToast`, `OverflowBanner`, `ActiveFiltersBar`) — requires careful prop-wiring + closure-semantics design and is deferred to PR-2d Phase 2. The orchestrator at 1100 lines is still over the 400-line internal target but comfortably under the 1500-line soft cap from issue #152 (no lint warning).
+  - 詳見 `docs/interactive/tools/tenant-manager/` directory + commit message + planning S#69
+
 - **`lint_jsx_babel.py` line-count guard（v2.8.0, lint, issue #152）** — codifies the lesson from PR #150 / S#67: `tenant-manager.jsx` accumulated 3 latent bugs over months because the 1691-line single file was too large for cross-checking `useState` declarations against `set*` references, hook ordering against early returns, and lifecycle paths in async branches. The guard adds a third validation pass to the existing JSX lint pipeline (alongside static patterns + Babel parse) — soft cap warns at **1500 lines**, hard cap fails at **2500 lines** under `--ci`. Soft-cap warnings stay non-fatal under default `--ci` (so commits aren't blocked during the v2.8.0 transition) and become fatal under `--ci --strict`. Hard-cap violations are always fatal under `--ci` regardless of `--strict`, mirroring how Babel parse errors are handled (severity-tier match: "file this big virtually guarantees latent bugs" = same blast radius as "file fails to parse"). Thresholds carry a `# TODO(v2.9.0): revisit` comment — they're conservative starting points (1500 picks today's single outlier `tenant-manager.jsx` at 1691 lines without flagging anyone else: next-largest is `operator-setup-wizard.jsx` at 1252; 2500 gives ~50% headroom over today's worst offender so it doesn't insta-fail current reality but blocks the next such offender). **32 new unit tests** in `tests/lint/test_lint_jsx_babel_linecount.py` — 15 pin `_run_line_count_check` constants + boundary semantics (`>` vs `>=` at both caps + trailing-newline counting matches `wc -l`) + 17 cover the `_compute_exit_code` severity matrix (8-state truth table over `--ci` × `--strict` × 4 failure categories), proving the contract. Includes verbatim acceptance criteria from issue #152 (2499 warns-but-passes; 2501 fails). The guard's actionable error message points to issue #153 (PR-2d file decomposition) so future hits know the pattern to apply. Pre-commit wiring: existing `jsx-babel-check` hook already runs `--ci`, so soft-cap warns + hard-cap blocks immediately on local commit; a manual-stage `--strict` hook can be added in a follow-up to also surface soft-cap on CI. Refactored `main()` to delegate severity computation to the pure `_compute_exit_code()` helper so the matrix is unit-testable without disk / Node.js.
 
 - **Phase .c — C-2 PR-2: tenant-manager.jsx API-first data source（v2.8.0）** — wires the C-1 server-side search endpoint (PR #148) into the Tenant Manager UI. The JSX now tries `GET /api/v1/tenants/search?page_size=500` first and falls back to `platform-data.json` (then `DEMO_TENANTS`) on any error. Addresses the planning §C-2 root cause for the 500+ tenant DOM-freeze problem: when the live API is reachable, the client never holds more than the page_size cap of summary records. UX preserves the existing client-side filter / search controls — they continue to operate over the visible set.
