+++ docs/interactive/tools/tenant-manager.jsx
@@ -5,623 +5,32 @@ audience: [platform-engineer, sre]
 version: v2.7.0
 lang: en
 related: [config-diff, playground, threshold-calculator, alert-simulator]
-dependencies: []
+dependencies: [
+  "tenant-manager/fixtures/demo-tenants.js",
+  "tenant-manager/styles.js",
+  "tenant-manager/utils/yaml-generators.js",
+  "tenant-manager/components/GroupSidebar.jsx"
+]
 ---
 
 import React, { useState, useMemo, useEffect, useCallback, useRef } from 'react';
 
 const t = window.__t || ((zh, en) => en);
 
-const DEMO_TENANTS = {
-  "prod-mariadb-01": {
-    environment: "production", region: "ap-northeast-1", tier: "tier-1",
-    domain: "finance", db_type: "mariadb",
-    rule_packs: ["mariadb", "kubernetes", "operational"],
-    owner: "team-dba-global", routing_channel: "slack:#dba-alerts",
-    operational_mode: "normal", metric_count: 8, last_config_commit: "abc1234",
-    tags: ["critical-path", "pci"], groups: ["production-dba"]
-  },
-  "prod-redis-01": {
-    environment: "production", region: "ap-northeast-1", tier: "tier-1",
-    domain: "cache", db_type: "redis",
-    rule_packs: ["redis", "kubernetes"],
-    owner: "team-platform", routing_channel: "pagerduty:dba-oncall",
-    operational_mode: "silent", metric_count: 5, last_config_commit: "abc1234",
-    tags: ["session-store"], groups: ["production-dba"]
-  },
-  "staging-pg-01": {
-    environment: "staging", region: "us-west-2", tier: "tier-2",
-    domain: "analytics", db_type: "postgresql",
-    rule_packs: ["postgresql", "jvm", "kubernetes"],
-    owner: "team-analytics", routing_channel: "slack:#staging-alerts",
-    operational_mode: "normal", metric_count: 6, last_config_commit: "def5678",
-    tags: [], groups: ["staging-all"]
-  },
-  "dev-mongodb-01": {
-    environment: "development", region: "us-west-2", tier: "tier-3",
-    domain: "mobile", db_type: "mongodb",
-    rule_packs: ["mongodb", "kubernetes"],
-    owner: "team-mobile", routing_channel: "email:mobile-dev@example.com",
-    operational_mode: "maintenance", metric_count: 3, last_config_commit: "ghi9012",
-    tags: ["experimental"], groups: []
-  },
-  "prod-kafka-01": {
-    environment: "production", region: "eu-west-1", tier: "tier-1",
-    domain: "streaming", db_type: "kafka",
-    rule_packs: ["kafka", "jvm", "kubernetes"],
-    owner: "team-streaming", routing_channel: "slack:#kafka-alerts",
-    operational_mode: "normal", metric_count: 7, last_config_commit: "abc1234",
-    tags: ["event-bus"], groups: ["production-dba"]
-  }
-};
+// PR-2d (#153) decomposition: pull dep symbols from window. Each dep
+// file (front-matter `dependencies: [...]` block above) self-registers
+// via `window.__X = X;` at its tail — same pattern self-service-portal
+// uses for its tab modules. This is the only safe way to share names
+// because jsx-loader evaluates deps via `(0, eval)(code)` (indirect
+// eval), where `const`/`let` declarations are block-scoped to the eval
+// frame and don't leak to the surrounding global scope.
+const DEMO_TENANTS = window.__DEMO_TENANTS;
+const DEMO_GROUPS = window.__DEMO_GROUPS;
+const styles = window.__styles;
+const generateMaintenanceYaml = window.__generateMaintenanceYaml;
+const generateSilentModeYaml = window.__generateSilentModeYaml;
+const GroupSidebar = window.__GroupSidebar;
 
-const DEMO_GROUPS = {
-  "production-dba": {
-    label: "Production DBA",
-    description: "All production database tenants managed by the DBA team",
-    members: ["prod-mariadb-01", "prod-redis-01", "prod-kafka-01"]
-  },
-  "staging-all": {
-    label: "All Staging",
-    description: "All staging environment tenants",
-    members: ["staging-pg-01"]
-  }
-};
-
-// ── Styles using design tokens (--da-*) from design-tokens.css ──
-// All hardcoded colors replaced with CSS variable references for theme support.
-// Style tokens (v2.7.0 Phase .a0 migration):
-// - Spacing / radius / font-size mapped to var(--da-*) tokens where exact match exists.
-// - Remaining literals are intentional: 6/11/14/18/28/48px are irregular values pending
-//   design-system decision (add bridge tokens vs. normalize to grid). See v2.7.0-planning.md §3.
-// - Layout constants (280/140/180/600/420/1400/1600px) stay literal — component-specific dimensions.
-const styles = {
-  container: {
-    minHeight: '100vh',
-    background: 'var(--da-color-bg)',
-    padding: 'var(--da-space-8)',
-  },
-  layout: {
-    display: 'grid',
-    gridTemplateColumns: '280px 1fr',
-    gap: 'var(--da-space-6)',
-    maxWidth: '1600px',
-    margin: '0 auto',
-  },
-  layoutNoSidebar: {
-    maxWidth: '1400px',
-    margin: '0 auto',
-  },
-  header: {
-    marginBottom: 'var(--da-space-6)',
-    maxWidth: '1600px',
-    margin: '0 auto var(--da-space-6)',
-  },
-  title: {
-    fontSize: 'var(--da-font-size-2xl)',
-    fontWeight: 'var(--da-font-weight-bold)',
-    color: 'var(--da-color-fg)',
-    marginBottom: 'var(--da-space-2)',
-  },
-  subtitle: {
-    fontSize: 'var(--da-font-size-base)',
-    color: 'var(--da-color-muted)',
-  },
-  authBanner: {
-    backgroundColor: 'var(--da-color-info-soft)',
-    border: '1px solid var(--da-color-accent)',
-    borderRadius: 'var(--da-radius-md)',
-    padding: 'var(--da-space-2) var(--da-space-4)',
-    marginBottom: 'var(--da-space-4)',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    fontSize: 'var(--da-font-size-sm-md)',
-    color: 'var(--da-color-info)',
-  },
-  sidebar: {
-    backgroundColor: 'var(--da-color-surface)',
-    border: '1px solid var(--da-color-surface-border)',
-    borderRadius: 'var(--da-radius-lg)',
-    padding: 'var(--da-space-4)',
-    boxShadow: 'var(--da-shadow-subtle)',
-    alignSelf: 'start',
-    position: 'sticky',
-    top: 'var(--da-space-8)',
-  },
-  sidebarTitle: {
-    fontSize: 'var(--da-font-size-md)',
-    fontWeight: 'var(--da-font-weight-bold)',
-    color: 'var(--da-color-fg)',
-    marginBottom: 'var(--da-space-3)',
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  groupItem: {
-    padding: '10px var(--da-space-3)',
-    borderRadius: 'var(--da-radius-md)',
-    cursor: 'pointer',
-    marginBottom: 'var(--da-space-1)',
-    fontSize: 'var(--da-font-size-sm-md)',
-    transition: 'background-color var(--da-transition-fast)',
-  },
-  groupItemActive: {
-    backgroundColor: 'var(--da-color-accent-soft)',
-    color: 'var(--da-color-accent)',
-    fontWeight: 'var(--da-font-weight-semibold)',
-  },
-  groupItemDefault: {
-    backgroundColor: 'transparent',
-    color: 'var(--da-color-fg)',
-  },
-  groupMemberCount: {
-    fontSize: 'var(--da-font-size-xs)',
-    color: 'var(--da-color-muted)',
-    marginLeft: 'var(--da-space-1)',
-  },
-  emptyState: {
-    textAlign: 'center',
-    padding: 'var(--da-space-8) var(--da-space-4)',
-    color: 'var(--da-color-muted)',
-    fontSize: 'var(--da-font-size-sm-md)',
-    lineHeight: 'var(--da-line-height-relaxed)',
-  },
-  statsBar: {
-    display: 'grid',
-    gridTemplateColumns: 'repeat(auto-fit, minmax(180px, 1fr))',
-    gap: 'var(--da-space-4)',
-    marginBottom: 'var(--da-space-6)',
-  },
-  statCard: {
-    backgroundColor: 'var(--da-color-surface)',
-    border: '1px solid var(--da-color-surface-border)',
-    borderRadius: 'var(--da-radius-lg)',
-    padding: 'var(--da-space-4)',
-    boxShadow: 'var(--da-shadow-subtle)',
-  },
-  statValue: {
-    fontSize: 'var(--da-font-size-xl)',
-    fontWeight: 'var(--da-font-weight-bold)',
-    color: 'var(--da-color-accent)',
-    marginBottom: 'var(--da-space-1)',
-  },
-  statLabel: {
-    fontSize: 'var(--da-font-size-xs)',
-    color: 'var(--da-color-muted)',
-    fontWeight: 'var(--da-font-weight-medium)',
-  },
-  controlsPanel: {
-    backgroundColor: 'var(--da-color-surface)',
-    border: '1px solid var(--da-color-surface-border)',
-    borderRadius: 'var(--da-radius-lg)',
-    padding: 'var(--da-space-4)',
-    marginBottom: 'var(--da-space-6)',
-    boxShadow: 'var(--da-shadow-subtle)',
-  },
-  searchInput: {
-    width: '100%',
-    padding: '10px var(--da-space-3)',
-    border: '1px solid var(--da-color-surface-border)',
-    borderRadius: 'var(--da-radius-md)',
-    fontSize: 'var(--da-font-size-base)',
-    marginBottom: 'var(--da-space-3)',
-    fontFamily: 'inherit',
-    boxSizing: 'border-box',
-    backgroundColor: 'var(--da-color-surface)',
-    color: 'var(--da-color-fg)',
-  },
-  filterRow: {
-    display: 'grid',
-    gridTemplateColumns: 'repeat(auto-fit, minmax(140px, 1fr))',
-    gap: 'var(--da-space-3)',
-    marginBottom: 'var(--da-space-3)',
-  },
-  filterSelect: {
-    padding: 'var(--da-space-2) var(--da-space-3)',
-    border: '1px solid var(--da-color-surface-border)',
-    borderRadius: 'var(--da-radius-md)',
-    fontSize: 'var(--da-font-size-sm-md)',
-    fontFamily: 'inherit',
-    backgroundColor: 'var(--da-color-surface)',
-    color: 'var(--da-color-fg)',
-  },
-  chipContainer: {
-    display: 'flex',
-    flexWrap: 'wrap',
-    gap: 'var(--da-space-2)',
-    marginBottom: 'var(--da-space-3)',
-  },
-  chip: {
-    display: 'inline-flex',
-    alignItems: 'center',
-    gap: '6px',
-    backgroundColor: 'var(--da-color-accent-soft)',
-    color: 'var(--da-color-accent)',
-    padding: '6px var(--da-space-3)',
-    borderRadius: 'var(--da-radius-pill)',
-    fontSize: 'var(--da-font-size-xs)',
-    fontWeight: 'var(--da-font-weight-medium)',
-  },
-  chipClose: {
-    cursor: 'pointer',
-    fontWeight: 'var(--da-font-weight-bold)',
-  },
-  buttonGroup: {
-    display: 'flex',
-    gap: 'var(--da-space-2)',
-    marginBottom: 'var(--da-space-3)',
-    flexWrap: 'wrap',
-  },
-  button: {
-    padding: 'var(--da-space-2) var(--da-space-4)',
-    backgroundColor: 'var(--da-color-accent)',
-    color: 'var(--da-color-accent-fg)',
-    border: 'none',
-    borderRadius: 'var(--da-radius-md)',
-    fontSize: 'var(--da-font-size-sm-md)',
-    fontWeight: 'var(--da-font-weight-medium)',
-    cursor: 'pointer',
-    transition: 'background-color var(--da-transition-base)',
-  },
-  buttonSecondary: {
-    backgroundColor: 'var(--da-color-tag-bg)',
-    color: 'var(--da-color-tag-fg)',
-    border: '1px solid var(--da-color-surface-border)',
-  },
-  buttonSmall: {
-    padding: 'var(--da-space-1) 10px',
-    fontSize: 'var(--da-font-size-xs)',
-  },
-  buttonDanger: {
-    backgroundColor: 'var(--da-color-error-soft)',
-    color: 'var(--da-color-error)',
-    border: '1px solid var(--da-color-error)',
-  },
-  buttonDisabled: {
-    opacity: 0.5,
-    cursor: 'not-allowed',
-  },
-  actionBar: {
-    backgroundColor: 'var(--da-color-info-soft)',
-    border: '1px solid var(--da-color-accent)',
-    borderRadius: 'var(--da-radius-md)',
-    padding: 'var(--da-space-3)',
-    marginBottom: 'var(--da-space-3)',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    flexWrap: 'wrap',
-    gap: 'var(--da-space-2)',
-  },
-  actionText: {
-    fontSize: 'var(--da-font-size-sm-md)',
-    color: 'var(--da-color-info)',
-    fontWeight: 'var(--da-font-weight-medium)',
-  },
-  grid: {
-    display: 'grid',
-    gridTemplateColumns: 'repeat(auto-fill, minmax(300px, 1fr))',
-    gap: 'var(--da-space-4)',
-    marginBottom: 'var(--da-space-6)',
-  },
-  card: {
-    backgroundColor: 'var(--da-color-surface)',
-    border: '1px solid var(--da-color-surface-border)',
-    borderRadius: 'var(--da-radius-lg)',
-    padding: 'var(--da-space-4)',
-    boxShadow: 'var(--da-shadow-subtle)',
-    transition: 'all var(--da-transition-base)',
-    cursor: 'pointer',
-    position: 'relative',
-  },
-  cardHover: {
-    boxShadow: 'var(--da-shadow-hover)',
-    transform: 'translateY(-2px)',
-  },
-  cardCheckbox: {
-    position: 'absolute',
-    top: 'var(--da-space-3)',
-    right: 'var(--da-space-3)',
-    width: 'var(--da-space-5)',
-    height: 'var(--da-space-5)',
-    cursor: 'pointer',
-  },
-  cardTitle: {
-    fontSize: '18px',
-    fontWeight: 'var(--da-font-weight-bold)',
-    color: 'var(--da-color-fg)',
-    marginBottom: 'var(--da-space-3)',
-    paddingRight: '28px',
-  },
-  badge: {
-    display: 'inline-block',
-    padding: 'var(--da-space-1) var(--da-space-2)',
-    borderRadius: 'var(--da-radius-sm)',
-    fontSize: 'var(--da-font-size-xs)',
-    fontWeight: 'var(--da-font-weight-semibold)',
-    marginRight: '6px',
-    marginBottom: 'var(--da-space-2)',
-  },
-  environmentBadge: {
-    production: { backgroundColor: 'var(--da-color-error-soft)', color: 'var(--da-color-error)' },
-    staging: { backgroundColor: 'var(--da-color-warning-soft)', color: 'var(--da-color-warning)' },
-    development: { backgroundColor: 'var(--da-color-success-soft)', color: 'var(--da-color-success)' },
-  },
-  tierBadge: {
-    'tier-1': { backgroundColor: 'var(--da-color-warning-soft)', color: 'var(--da-color-warning)' },
-    'tier-2': { backgroundColor: 'var(--da-color-tag-bg)', color: 'var(--da-color-tag-fg)' },
-    'tier-3': { backgroundColor: 'var(--da-color-tag-bg)', color: 'var(--da-color-tag-fg)' },
-  },
-  pills: {
-    display: 'flex',
-    flexWrap: 'wrap',
-    gap: 'var(--da-space-1)',
-    marginBottom: 'var(--da-space-3)',
-  },
-  pill: {
-    backgroundColor: 'var(--da-color-tag-bg)',
-    color: 'var(--da-color-tag-fg)',
-    padding: 'var(--da-space-1) var(--da-space-2)',
-    borderRadius: 'var(--da-radius-pill)',
-    fontSize: 'var(--da-font-size-xs)',
-    fontWeight: 'var(--da-font-weight-medium)',
-  },
-  row: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    padding: 'var(--da-space-2) 0',
-    borderTop: '1px solid var(--da-color-tag-bg)',
-    fontSize: 'var(--da-font-size-sm-md)',
-  },
-  rowLabel: {
-    color: 'var(--da-color-muted)',
-    fontWeight: 'var(--da-font-weight-medium)',
-  },
-  rowValue: {
-    color: 'var(--da-color-fg)',
-    fontWeight: 'var(--da-font-weight-semibold)',
-  },
-  modeIndicator: {
-    display: 'inline-block',
-    width: 'var(--da-space-2)',
-    height: 'var(--da-space-2)',
-    borderRadius: 'var(--da-radius-full)',
-    marginRight: 'var(--da-space-1)',
-  },
-  modal: {
-    position: 'fixed',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    backgroundColor: 'var(--da-color-modal-backdrop)',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    zIndex: 1000,
-  },
-  modalContent: {
-    backgroundColor: 'var(--da-color-surface)',
-    borderRadius: 'var(--da-radius-lg)',
-    padding: 'var(--da-space-6)',
-    maxWidth: '600px',
-    width: '90%',
-    maxHeight: '80vh',
-    overflow: 'auto',
-    boxShadow: 'var(--da-shadow-modal)',
-  },
-  modalTitle: {
-    fontSize: '18px',
-    fontWeight: 'var(--da-font-weight-bold)',
-    color: 'var(--da-color-fg)',
-    marginBottom: 'var(--da-space-4)',
-  },
-  codeBlock: {
-    backgroundColor: 'var(--da-color-hero-bg)',
-    color: 'var(--da-color-hero-fg)',
-    padding: 'var(--da-space-3)',
-    borderRadius: 'var(--da-radius-md)',
-    fontFamily: 'var(--da-font-mono)',
-    fontSize: 'var(--da-font-size-xs)',
-    lineHeight: '1.4',
-    marginBottom: 'var(--da-space-3)',
-    overflow: 'auto',
-    whiteSpace: 'pre-wrap',
-  },
-  buttonGroup2: {
-    display: 'flex',
-    gap: 'var(--da-space-2)',
-    marginTop: 'var(--da-space-4)',
-  },
-  tooltip: {
-    position: 'relative',
-    display: 'inline-block',
-  },
-  tooltipText: {
-    fontSize: 'var(--da-font-size-xs)',
-    color: 'var(--da-color-muted)',
-    fontStyle: 'italic',
-    marginTop: 'var(--da-space-1)',
-  },
-  inputField: {
-    width: '100%',
-    padding: 'var(--da-space-2) var(--da-space-3)',
-    border: '1px solid var(--da-color-surface-border)',
-    borderRadius: 'var(--da-radius-md)',
-    fontSize: 'var(--da-font-size-sm-md)',
-    fontFamily: 'inherit',
-    marginBottom: 'var(--da-space-2)',
-    boxSizing: 'border-box',
-    backgroundColor: 'var(--da-color-surface)',
-    color: 'var(--da-color-fg)',
-  },
-  formLabel: {
-    display: 'block',
-    fontSize: 'var(--da-font-size-xs)',
-    fontWeight: 'var(--da-font-weight-semibold)',
-    color: 'var(--da-color-tag-fg)',
-    marginBottom: 'var(--da-space-1)',
-  },
-};
-
-function generateMaintenanceYaml(tenants) {
-  const lines = [];
-  lines.push('apiVersion: v1');
-  lines.push('kind: ConfigMap');
-  lines.push('metadata:');
-  lines.push('  name: tenant-operational-modes');
-  lines.push('  namespace: monitoring');
-  lines.push('data:');
-  tenants.forEach(name => {
-    lines.push(`  ${name}_maintenance: |`);
-    lines.push(`    mode: maintenance`);
-    lines.push(`    reason: "Scheduled maintenance"`);
-    lines.push(`    expires: "2026-04-05T00:00:00Z"`);
-  });
-  return lines.join('\n');
-}
-
-function generateSilentModeYaml(tenants) {
-  const lines = [];
-  lines.push('apiVersion: v1');
-  lines.push('kind: ConfigMap');
-  lines.push('metadata:');
-  lines.push('  name: tenant-operational-modes');
-  lines.push('  namespace: monitoring');
-  lines.push('data:');
-  tenants.forEach(name => {
-    lines.push(`  ${name}_silent: |`);
-    lines.push(`    mode: silent`);
-    lines.push(`    reason: "Under investigation"`);
-    lines.push(`    expires: "2026-04-04T12:00:00Z"`);
-  });
-  return lines.join('\n');
-}
-
-// ─── Group Sidebar Component ─────────────────────────────────────────────────
-function GroupSidebar({ groups, activeGroupId, onSelectGroup, onCreateGroup, onDeleteGroup, canWrite }) {
-  const [showCreate, setShowCreate] = useState(false);
-  const [newGroupId, setNewGroupId] = useState('');
-  const [newGroupLabel, setNewGroupLabel] = useState('');
-
-  const handleCreate = () => {
-    if (newGroupId && newGroupLabel) {
-      onCreateGroup({ id: newGroupId.toLowerCase().replace(/\s+/g, '-'), label: newGroupLabel });
-      setNewGroupId('');
-      setNewGroupLabel('');
-      setShowCreate(false);
-    }
-  };
-
-  const groupEntries = Object.entries(groups);
-
-  return (
-    <div style={styles.sidebar} role="complementary" aria-label={t('群組管理側欄', 'Group management sidebar')} aria-live="polite" aria-atomic="true">
-      <div style={styles.sidebarTitle}>
-        <span>{t('群組', 'Groups')} ({groupEntries.length})</span>
-        {canWrite && (
-          <button
-            onClick={() => setShowCreate(!showCreate)}
-            style={{ ...styles.button, ...styles.buttonSmall }}
-            title={t('建立新群組 — 用於批量操作和分組篩選', 'Create new group — for batch operations and group filtering')}
-            aria-label={showCreate ? t('取消建立群組', 'Cancel create group') : t('建立新群組', 'Create new group')}
-            aria-expanded={showCreate}
-          >
-            {showCreate ? '−' : '+'}
-          </button>
-        )}
-      </div>
-
-      {showCreate && (
-        <div style={{ marginBottom: 'var(--da-space-3)', padding: 'var(--da-space-2)', backgroundColor: 'var(--da-color-bg)', borderRadius: 'var(--da-radius-md)' }}>
-          <label htmlFor="new-group-id" style={styles.formLabel}>{t('群組 ID', 'Group ID')}</label>
-          <input
-            id="new-group-id"
-            style={styles.inputField}
-            placeholder="e.g., production-dba"
-            value={newGroupId}
-            onChange={e => setNewGroupId(e.target.value)}
-          />
-          <label htmlFor="new-group-label" style={styles.formLabel}>{t('顯示名稱', 'Display Label')}</label>
-          <input
-            id="new-group-label"
-            style={styles.inputField}
-            placeholder="e.g., Production DBA"
-            value={newGroupLabel}
-            onChange={e => setNewGroupLabel(e.target.value)}
-          />
-          <button onClick={handleCreate} style={{ ...styles.button, ...styles.buttonSmall, width: '100%' }}>
-            {t('建立群組', 'Create Group')}
-          </button>
-        </div>
-      )}
-
-      {/* "All Tenants" entry */}
-      <div
-        style={{
-          ...styles.groupItem,
-          ...(activeGroupId === null ? styles.groupItemActive : styles.groupItemDefault),
-        }}
-        role="button"
-        tabIndex={0}
-        aria-label={t('所有租戶', 'All Tenants')}
-        aria-pressed={activeGroupId === null}
-        onClick={() => onSelectGroup(null)}
-        onKeyDown={e => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); onSelectGroup(null); } }}
-      >
-        {t('所有租戶', 'All Tenants')}
-      </div>
-
-      {groupEntries.length === 0 && (
-        <div style={styles.emptyState}>
-          <div style={{ fontSize: 'var(--da-font-size-xl)', marginBottom: 'var(--da-space-2)' }} aria-hidden="true">📁</div>
-          <div>{t('尚無群組', 'No groups yet')}</div>
-          <div style={{ fontSize: '11px', marginTop: 'var(--da-space-1)' }}>
-            {t('建立群組以批量管理租戶', 'Create a group to batch-manage tenants')}
-          </div>
-        </div>
-      )}
-
-      {groupEntries.map(([id, group]) => (
-        <div
-          key={id}
-          style={{
-            ...styles.groupItem,
-            ...(activeGroupId === id ? styles.groupItemActive : styles.groupItemDefault),
-            display: 'flex',
-            justifyContent: 'space-between',
-            alignItems: 'center',
-          }}
-          role="button"
-          tabIndex={0}
-          aria-label={`${t('選擇群組', 'Select group')}: ${group.label || id}`}
-          aria-pressed={activeGroupId === id}
-          onClick={() => onSelectGroup(id)}
-          onKeyDown={e => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); onSelectGroup(id); } }}
-        >
-          <span>
-            {group.label || id}
-            <span style={styles.groupMemberCount}>({(group.members || []).length})</span>
-          </span>
-          {canWrite && (
-            <button
-              style={{ ...styles.chipClose, fontSize: '14px', color: 'var(--da-color-muted)', background: 'none', border: 'none', cursor: 'pointer', padding: '2px 4px' }}
-              title={t('刪除此群組', 'Delete this group')}
-              aria-label={`${t('刪除群組', 'Delete group')}: ${group.label || id}`}
-              onClick={(e) => { e.stopPropagation(); onDeleteGroup(id); }}
-            >
-              ✕
-            </button>
-          )}
-        </div>
-      ))}
-    </div>
-  );
-}
-
-// ─── Main Component ──────────────────────────────────────────────────────────
 export default function TenantManager() {
   const [tenants, setTenants] = useState({});
   const [groups, setGroups] = useState({});
